@@ -4,28 +4,37 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "build": "vue-cli-service build",
+    "deploy": "firebase deploy"
   },
   "dependencies": {
-    "chart.js": "^3.8.0",
-    "core-js": "^3.8.3",
-    "firebase": "^10.0.0",
+    "@vee-validate/yup": "^4.11.7",
+    "chart.js": "^4.4.0",
+    "core-js": "^3.32.2",
+    "firebase": "^10.4.0",
     "lodash": "^4.17.21",
-    "materialize-css": "^1.0.0-rc.2",
-    "vue": "^2.6.14",
-    "vue-chartjs": "^4.0.0",
-    "vue-meta": "^2.4.0",
-    "vue-router": "^3.5.1",
-    "vuejs-paginate": "^2.1.0",
-    "vuelidate": "^0.7.7",
-    "vuex": "^3.6.2"
+    "materialize-css": "^1.0.0",
+    "vee-validate": "^4.11.7",
+    "vue": "^3.3.4",
+    "vue-chartjs": "^5.2.0",
+    "vue-meta": "^3.0.0-alpha.10",
+    "vue-router": "^4.2.5",
+    "vuejs-paginate-next": "^1.0.2",
+    "vuex": "^4.1.0",
+    "yup": "^1.3.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "@babel/core": "^7.23.0",
+    "@babel/eslint-parser": "^7.22.15",
+    "@vue/cli-plugin-babel": "~5.0.8",
+    "@vue/cli-plugin-eslint": "~5.0.8",
+    "@vue/cli-service": "~5.0.8",
+    "@vue/eslint-config-standard": "^8.0.1",
+    "eslint": "^8.50.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-perfectionist": "^2.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.17.0"
   }
 }
