@@ -1,54 +1,42 @@
-import Vue from 'vue'
-import VueMeta from 'vue-meta'
-import Vuelidate from 'vuelidate'
 import { initializeApp } from 'firebase/app'
 import { getAuth, onAuthStateChanged } from 'firebase/auth'
-import Paginate from 'vuejs-paginate'
+import 'materialize-css/dist/css/materialize.min.css'
+import 'materialize-css/dist/js/materialize.min.js'
+import { createApp } from 'vue'
+import { createMetaManager } from 'vue-meta'
+import Paginate from 'vuejs-paginate-next'
 
 import App from './App.vue'
+import './assets/index.css'
+import AppLoader from './components/app/AppLoader'
+import tooltipDirective from './directives/tooltip'
+import messagePlugin from './plugins/message'
 import router from './router'
 import store from './store'
-import dateFilter from './filters/date.filter'
-import currencyFilter from './filters/currency.filter'
-import localizeFilter from './filters/localize.filter'
-import messagePlugin from './utils/message.plugin'
-import titlePlugin from './utils/title.plugin'
-import tooltipDirective from './directives/tooltip.directive'
-import Loader from './components/app/Loader'
-
-import 'materialize-css/dist/js/materialize.min.js'
-
-Vue.config.productionTip = false
-
-Vue.use(VueMeta)
-Vue.use(Vuelidate)
-Vue.use(messagePlugin)
-Vue.use(titlePlugin)
-Vue.filter('date', dateFilter)
-Vue.filter('currency', currencyFilter)
-Vue.filter('localize', localizeFilter)
-Vue.directive('tooltip', tooltipDirective)
-Vue.component('Loader', Loader)
-Vue.component('Paginate', Paginate)
 
 initializeApp({
   apiKey: 'AIzaSyC2DiliMxVbxN5OsYUwc5d9Pos7A2yH9mw',
-  authDomain: 'vue-crm-93feb.firebaseapp.com',
-  projectId: 'vue-crm-93feb',
-  storageBucket: 'vue-crm-93feb.appspot.com',
-  messagingSenderId: '207435020724',
   appId: '1:207435020724:web:96e6c441679d83cc6adae1',
-  databaseURL: 'https://vue-crm-93feb-default-rtdb.europe-west1.firebasedatabase.app/'
+  authDomain: 'vue-crm-93feb.firebaseapp.com',
+  databaseURL: 'https://vue-crm-93feb-default-rtdb.europe-west1.firebasedatabase.app/',
+  messagingSenderId: '207435020724',
+  projectId: 'vue-crm-93feb',
+  storageBucket: 'vue-crm-93feb.appspot.com'
 })
 
 let app
 const auth = getAuth()
-onAuthStateChanged(auth, () => {
+onAuthStateChanged(auth, async () => {
   if (!app) {
-    app = new Vue({
-      router,
-      store,
-      render: h => h(App)
-    }).$mount('#app')
+    app = createApp(App)
+    app.use(store)
+    app.use(router)
+    app.use(createMetaManager())
+    app.use(messagePlugin)
+    app.directive('tooltip', tooltipDirective)
+    app.component('AppLoader', AppLoader)
+    app.component('AppPagination', Paginate)
+    await router.isReady()
+    app.mount('#app')
   }
 })
