@@ -1,89 +1,87 @@
 <template>
   <div>
     <div class="page-title">
-      <h3>{{ 'Menu_Planning' | localize }}</h3>
-      <h4>{{ info.bill | currency }}</h4>
+      <h3>{{ localize('Menu_Planning') }}</h3>
+      <h4>{{ currencyFormat(info.bill) }}</h4>
     </div>
 
-    <Loader v-if="!ready" />
+    <app-loader v-if="!ready"></app-loader>
 
-    <p
-      class="center"
-      v-else-if="!categories.length"
-    >
-      {{ 'NoCategories' | localize }}.
-      <RouterLink to="/categories">{{ 'AddFirst' | localize }}</RouterLink>
-    </p>
-
-    <section v-else>
+    <section v-else-if="categories.length">
       <div
-        v-for="cat in categories"
         :key="cat.id"
+        v-for="cat in categories"
         v-tooltip.noloc="{ html: cat.tooltip }"
       >
         <p>
           <strong>{{ cat.title }}:</strong>
-          {{ cat.balance | currency }} {{ 'Of' | localize }} {{ cat.limit | currency }}
+          {{ currencyFormat(cat.balance) }} {{ localize('Of') }} {{ currencyFormat(cat.limit) }}
         </p>
         <div class="progress">
           <div
-            class="determinate"
             :class="[cat.progressColor]"
             :style="{ width: `${cat.progressPercent}%` }"
+            class="determinate"
           ></div>
         </div>
       </div>
     </section>
+
+    <p
+      class="center"
+      v-else
+    >
+      {{ localize('NoCategories') }}.
+      <router-link to="/categories">{{ localize('AddFirst') }}</router-link>
+    </p>
   </div>
 </template>
 
-<script>
-import { mapGetters } from 'vuex'
-import currencyFilter from '@/filters/currency.filter'
-import localizeFilter from '@/filters/localize.filter'
-import isReady from '@/helpers/isReady'
+<script setup>
+import { computed, onMounted } from 'vue'
+import { useMeta } from 'vue-meta'
+import { useStore } from 'vuex'
 
-export default {
-  name: 'planning',
-  metaInfo() {
+import isReady from '../helpers/isReady'
+import currencyFormat from '../utils/currencyFormat'
+import localize from '../utils/localize'
+
+useMeta({ title: 'Menu_Planning' })
+
+const store = useStore()
+const info = computed(() => store.getters['info/info'])
+const records = computed(() => store.getters['record/records'])
+const categories = computed(() => {
+  return store.getters['category/categories'].map(c => {
+    const balance = records.value
+      .filter(r => r.categoryId === c.id)
+      .reduce((total, r) => {
+        return r.type === 'outcome' ? total + r.amount : total - r.amount
+      }, 0)
+    const percent = (100 * balance) / c.limit
+    const progressPercent = Math.min(percent, 100)
+    const progressColor = percent < 60 ? 'green' : percent < 100 ? 'yellow' : 'red'
+
+    const tooltipValue = c.limit - balance
+    const tooltip = `${tooltipValue < 0 ? localize('MoreThan') : localize('Stayed')} ${currencyFormat(
+      Math.abs(tooltipValue)
+    )}`
+
     return {
-      title: this.$title('Menu_Planning')
+      ...c,
+      balance,
+      progressColor,
+      progressPercent,
+      tooltip
     }
-  },
-  computed: {
-    ...mapGetters(['info', 'records']),
-    categories() {
-      return this.$store.getters.categories.map(c => {
-        const balance = this.records
-          .filter(r => r.categoryId === c.id)
-          .reduce((total, r) => {
-            return r.type === 'outcome' ? total + r.amount : total - r.amount
-          }, 0)
-        const percent = (100 * balance) / c.limit
-        const progressPercent = Math.min(percent, 100)
-        const progressColor = percent < 60 ? 'green' : percent < 100 ? 'yellow' : 'red'
+  })
+})
+const ready = computed(() => {
+  return isReady(store.getters['category/categoriesReady'], store.getters['record/recordsReady'])
+})
 
-        const tooltipValue = c.limit - balance
-        const tooltip = `${tooltipValue < 0 ? localizeFilter('MoreThan') : localizeFilter('Stayed')} ${currencyFilter(
-          Math.abs(tooltipValue)
-        )}`
-
-        return {
-          ...c,
-          progressPercent,
-          progressColor,
-          balance,
-          tooltip
-        }
-      })
-    },
-    ready() {
-      return isReady(this.$store.getters.categoriesReady, this.$store.getters.recordsReady)
-    }
-  },
-  async mounted() {
-    await this.$store.dispatch('fetchCategories')
-    await this.$store.dispatch('fetchRecords')
-  }
-}
+onMounted(async () => {
+  await store.dispatch('category/fetchCategories')
+  await store.dispatch('record/fetchRecords')
+})
 </script>
