@@ -1,7 +1,7 @@
 export default {
-  logout: 'Logout',
-  login: 'FirstLogin',
+  'auth/email-already-in-use': 'EmailInUse',
   'auth/user-not-found': 'NoUserWithEmail',
   'auth/wrong-password': 'WrongPassword',
-  'auth/email-already-in-use': 'EmailInUse'
+  login: 'FirstLogin',
+  logout: 'Logout'
 }
