@@ -1,58 +1,55 @@
 <template>
   <div>
     <div class="page-title">
-      <h3>{{ 'Bill' | localize }}</h3>
+      <h3>{{ localize('Bill') }}</h3>
 
       <button
-        class="btn waves-effect waves-light btn-small"
         @click="refresh"
+        class="btn waves-effect waves-light btn-small"
       >
         <i class="material-icons">refresh</i>
       </button>
     </div>
 
-    <Loader v-if="loading" />
+    <app-loader v-if="loading"></app-loader>
 
     <div
       class="row"
       v-else
     >
-      <HomeBill :rates="currency.rates" />
+      <home-bill :rates="currency.rates"></home-bill>
 
-      <HomeCurrency
-        :rates="currency.rates"
+      <home-currency
         :date="currency.date"
-      />
+        :rates="currency.rates"
+      ></home-currency>
     </div>
   </div>
 </template>
 
-<script>
-import HomeBill from '@/components/HomeBill.vue'
-import HomeCurrency from '@/components/HomeCurrency.vue'
+<script setup>
+import { onMounted, ref } from 'vue'
+import { useMeta } from 'vue-meta'
+import { useStore } from 'vuex'
 
-export default {
-  name: 'home',
-  metaInfo() {
-    return {
-      title: this.$title('Menu_Bill')
-    }
-  },
-  data: () => ({
-    loading: true,
-    currency: null
-  }),
-  async mounted() {
-    this.currency = await this.$store.dispatch('fetchCurrency')
-    this.loading = false
-  },
-  methods: {
-    async refresh() {
-      this.loading = true
-      this.currency = await this.$store.dispatch('fetchCurrency')
-      this.loading = false
-    }
-  },
-  components: { HomeBill, HomeCurrency }
+import HomeBill from '../components/home/HomeBill.vue'
+import HomeCurrency from '../components/home/HomeCurrency.vue'
+import localize from '../utils/localize'
+
+useMeta({ title: 'Menu_Bill' })
+
+const store = useStore()
+const currency = ref(null)
+const loading = ref(true)
+
+const refresh = async () => {
+  loading.value = true
+  currency.value = await store.dispatch('fetchCurrency')
+  loading.value = false
 }
+
+onMounted(async () => {
+  currency.value = await store.dispatch('fetchCurrency')
+  loading.value = false
+})
 </script>
